Quite a few changes:
First was to reduce the assembly to remove anything that has an address associated with it. This 
improves the diff capability and homogenizes the assembly regardless of address. 

Second was to add a count to the assembly sequence, this is used to apply the changes easier. The 
previous method worked but was clunky and unreliable. The idea is that we walk through the change
line by line. 

Last was to incorporate diff_match_patch (dmp) from:
http://code.google.com/p/google-diff-match-patch/
This was the best solution for difference that worked better than Python Diff, and 
PatienceSequenceMatcher. PatienceSequenceMatcher got close but did not handle the line changes
as well as dmp

TODO: 
Fix the diff output as the one from dmp sucks
Add local names to the instructions as it is currently dropped
Clean up diff and add verbose logging